% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_ball_sampling.R
\name{chisq_divergence}
\alias{chisq_divergence}
\title{Compute Chi-Squared Divergence Between Distributions (Internal)}
\usage{
chisq_divergence(Q, P0)
}
\arguments{
\item{Q}{Numeric vector. Probability distribution.}

\item{P0}{Numeric vector. Baseline probability distribution (full support).}
}
\value{
Numeric. `sum((Q - P0)^2 / P0)`.
}
\description{
Compute Chi-Squared Divergence Between Distributions (Internal)
}
\keyword{internal}
