% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_influence_function.R
\name{compute_functional_from_effects}
\alias{compute_functional_from_effects}
\title{Compute functional from treatment effect pairs}
\usage{
compute_functional_from_effects(
  delta_s_vec,
  delta_y_vec,
  functional_type,
  epsilon_s,
  epsilon_y
)
}
\description{
Compute functional from treatment effect pairs
}
\keyword{internal}
