% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_level_minimax.R
\name{compute_functional_from_effects_minimax}
\alias{compute_functional_from_effects_minimax}
\title{Compute Functional from Treatment Effects (Minimax Version)}
\usage{
compute_functional_from_effects_minimax(
  effects,
  functional_type,
  epsilon_s,
  epsilon_y,
  delta_s_value
)
}
\arguments{
\item{effects}{Matrix of treatment effects (M x 2): [delta_s, delta_y]}

\item{functional_type}{Character: type of functional}

\item{epsilon_s}{Threshold for probability functionals}

\item{epsilon_y}{Threshold for probability functionals}

\item{delta_s_value}{Conditioning value for conditional_mean}
}
\value{
Scalar minimax functional value
}
\description{
Computes minimax functional value from distribution of treatment effects.
}
\details{
For minimax inference, we want the MINIMUM value of the functional
over the TV-ball. For correlation, we take the minimum correlation
over bootstrap subsamples to get a robust estimate.
}
\keyword{internal}
