% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_inference.R
\name{compute_summary_stats_nested}
\alias{compute_summary_stats_nested}
\title{Compute summary statistics for nested bootstrap}
\usage{
compute_summary_stats_nested(mean_estimates, all_samples)
}
\arguments{
\item{mean_estimates}{Numeric vector. M estimates of E[phi(F_lambda)]}

\item{all_samples}{Numeric vector. All M*B samples of phi(Q)}
}
\description{
Computes both CI for E[phi] and prediction interval for phi(Q).
}
\keyword{internal}
