% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_minimax.R
\name{construct_search_grid}
\alias{construct_search_grid}
\title{Construct search grid for minimax inference}
\usage{
construct_search_grid(
  n,
  dirichlet_alpha_range,
  n_dirichlet_grid,
  include_vertices,
  max_vertices
)
}
\arguments{
\item{n}{Sample size}

\item{dirichlet_alpha_range}{Range of alpha values}

\item{n_dirichlet_grid}{Number of Dirichlet grid points}

\item{include_vertices}{Include vertex distributions?}

\item{max_vertices}{Maximum number of vertices}
}
\value{
Tibble with mu_type, alpha, vertex_id columns
}
\description{
Builds grid of innovation distributions to search over.
}
\keyword{internal}
