% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{discretize_kmeans}
\alias{discretize_kmeans}
\title{K-Means Clustering Discretization}
\usage{
discretize_kmeans(data, covariate_cols, k = 9)
}
\arguments{
\item{data}{Data frame with covariates}

\item{covariate_cols}{Character vector of covariate column names}

\item{k}{Number of clusters (types)}
}
\value{
Integer vector of cluster assignments (length nrow(data))
}
\description{
Discretizes observations via k-means clustering on covariates.
}
\details{
Performs k-means clustering on scaled covariates. This creates bins
that are compact in covariate space (Euclidean distance). Unlike
quantile binning, k-means adapts to covariate distribution.
}
\keyword{internal}
