% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein_minimax_IF_inference.R
\name{estimate_dual_fold_wasserstein}
\alias{estimate_dual_fold_wasserstein}
\title{Estimate Wasserstein Dual on Fold (Internal)}
\usage{
estimate_dual_fold_wasserstein(test_data, h_hat, covariates, gamma, tau)
}
\arguments{
\item{test_data}{Test fold}

\item{h_hat}{Concordance estimates}

\item{covariates}{Covariate names}

\item{gamma}{Penalty}

\item{tau}{Temperature}
}
\value{
Numeric: dual estimate
}
\description{
Estimate Wasserstein Dual on Fold (Internal)
}
\keyword{internal}
