% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_minimax.R
\name{evaluate_phi_at_grid_point}
\alias{evaluate_phi_at_grid_point}
\title{Evaluate phi at a specific grid point}
\usage{
evaluate_phi_at_grid_point(
  grid_row,
  current_data,
  lambda,
  n_innovations,
  functional_type,
  epsilon_s,
  epsilon_y,
  delta_s_value
)
}
\arguments{
\item{grid_row}{Single-row tibble from search_grid}

\item{current_data}{Data frame}

\item{lambda}{Perturbation parameter}

\item{n_innovations}{Number of MC innovations}

\item{functional_type}{Type of functional}

\item{epsilon_s}{Threshold for probability}

\item{epsilon_y}{Threshold for probability}

\item{delta_s_value}{Conditioning value for conditional_mean}
}
\value{
Scalar phi value
}
\description{
Computes phi(F_lambda) for a given innovation distribution specification.
}
\keyword{internal}
