% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_level_minimax.R
\name{fit_and_predict_tau}
\alias{fit_and_predict_tau}
\title{Fit and Predict Treatment Effect (Internal Helper)}
\usage{
fit_and_predict_tau(
  data_train,
  data_test,
  outcome,
  covariates,
  method,
  bandwidth
)
}
\description{
Fit and Predict Treatment Effect (Internal Helper)
}
\keyword{internal}
