% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_level_minimax.R
\name{fit_tau_model}
\alias{fit_tau_model}
\title{Fit Treatment Effect Model (for returning model object)}
\usage{
fit_tau_model(data, outcome, covariates, method, bandwidth)
}
\description{
Fit Treatment Effect Model (for returning model object)
}
\keyword{internal}
