% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimax_wrappers.R
\name{minimax_wrappers}
\alias{minimax_wrappers}
\title{User-Facing Minimax Wrapper Functions}
\description{
Simple wrappers for minimax inference that take type-level treatment effects
as input. Used by simulation studies.
}
