% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_ball_sampling.R
\name{sample_simplex_direction}
\alias{sample_simplex_direction}
\title{Sample Uniform Direction on Probability Simplex Tangent Space (Internal)}
\usage{
sample_simplex_direction(K)
}
\arguments{
\item{K}{Integer. Dimension of simplex (number of types).}
}
\value{
Numeric vector of length K with sum = 0 and norm = 1.
}
\description{
Sample Uniform Direction on Probability Simplex Tangent Space (Internal)
}
\keyword{internal}
