% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_ball_sampling.R
\name{tv_distance}
\alias{tv_distance}
\title{Compute TV Distance Between Distributions (Internal)}
\usage{
tv_distance(Q, P0)
}
\arguments{
\item{Q}{Numeric vector. Probability distribution.}

\item{P0}{Numeric vector. Baseline probability distribution.}
}
\value{
Numeric. TV distance = (1/2) * sum(|Q - P0|).
}
\description{
Compute TV Distance Between Distributions (Internal)
}
\keyword{internal}
