% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_influence_function.R
\name{gradient_correlation_analytical}
\alias{gradient_correlation_analytical}
\title{Analytical gradient for correlation functional}
\usage{
gradient_correlation_analytical(
  delta_s,
  delta_y,
  lambda,
  alpha,
  n_innovations,
  data
)
}
\description{
For correlation, we can compute the gradient analytically
}
\keyword{internal}
